Refatora service para retornar uma lista com a quantidade de transaçoes por banco

+++ src/main/java/com/cielo/dto/PercentPayment.java
@@ -14,7 +14,7 @@ import java.util.List;
 @Builder
 public class PercentPayment {
 
-    private Integer percentPaid;
+    private Double percentPaid;
     private Integer totalPayments;
     private Integer paymentsPaid;
     private List<ControleLancamento> controleLancamentoList;
